fix(local-dev): first-run paper cuts on Windows + tsx watch
Five small fixes from the May 19 local deployment review
(reports/deft-local-deployment-and-platform-review.pdf):

1. apps/api/src/index.ts → splits into index.ts (exports `app`) + new
   server.ts (binds the port). The old `if (isMain) serve()` guard
   used a brittle process.argv[1] string-endsWith check that silently
   disabled `pnpm dev` under tsx watch (argv[1] becomes the tsx CLI,
   not our source file). Splitting removes the guard entirely — tests
   import `app` from index.ts without binding a port, dev/start/build
   point at server.ts. Dockerfile CMD updated to match.

2. apps/api/src/index.ts CORS now accepts an array of allowed origins
   so both `localhost:3000` and `127.0.0.1:3000` work in development.
   Windows browsers sometimes default to 127.0.0.1, which failed
   preflight against the single-origin config.

3. apps/web/next.config.ts adds `allowedDevOrigins: ['localhost',
   '127.0.0.1']` for the matching Next-side check.

4. apps/api/src/routes/messages.ts — the Defty enqueue and the BYOA
   `agent-employee-message` enqueue both gated on env.ANTHROPIC_API_KEY,
   so org-stored Anthropic keys (the documented happy path in
   /settings/ai) never triggered an agent reply. Both now call
   resolveAnthropicApiKey(user.org_id), which is what the streaming
   agent path already uses. Resolved once per request and reused.

5. packages/db/scripts/apply-extras.ts loads the repo-root .env via
   dotenv. `pnpm db:push-full` chains drizzle-kit push (which loads
   .env through drizzle.config.ts) with this script (which read
   process.env.DATABASE_URL directly), so fresh-clone runs pushed the
   schema and then died at the extras step. Both halves now see the
   same env.

Findings #2, #3, #6 from the report (compose Postgres port,
DISABLE_ENV_AI_FALLBACK, llm.ts default routes) are out of scope here —
the compose port has a deliberate security-comment defense, and the
fallback / model-routing UX is a bigger product call.
